refactor db.js into 3 files (models, init, methods)
